requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
abseil-cpp=20211102.0=hd4dd3e8_0
anyio=4.2.0=pypi_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow-cpp=11.0.0=h374c478_2
asttokens=2.0.5=pyhd3eb1b0_0
async-lru=2.0.4=pypi_0
attrs=23.1.0=pypi_0
aws-c-common=0.6.8=h5eee18b_1
aws-c-event-stream=0.1.6=h6a678d5_6
aws-checksums=0.1.11=h5eee18b_2
aws-sdk-cpp=1.8.185=h721c034_1
babel=2.11.0=pypi_0
beautifulsoup4=4.12.2=pypi_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
boost-cpp=1.82.0=hdb19cb5_2
brotli=1.0.9=pypi_0
brotli-python=1.0.9=py311h6a678d5_7
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=h5eee18b_0
ca-certificates=2023.12.12=h06a4308_0
certifi=2023.11.17=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
comm=0.1.2=pypi_0
configparser=5.0.2=pyhd3eb1b0_0
cryptography=41.0.7=pypi_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
debugpy=1.6.7=pypi_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
executing=0.8.3=pyhd3eb1b0_0
fastjsonschema=2.16.2=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=pypi_0
freetype=2.12.1=h4a9f257_0
future=0.18.3=pypi_0
gflags=2.2.2=he6710b0_0
giflib=5.2.1=h5eee18b_3
glog=0.5.0=h2531618_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=pypi_0
gnutls=3.6.15=he1e5248_0
grpc-cpp=1.48.2=he1ff14a_1
icu=73.1=h6a678d5_0
idna=3.4=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.28.0=pypi_0
ipython=8.20.0=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
jpeg=9e=h5eee18b_1
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.19.2=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
jupyter-client=8.6.0=pypi_0
jupyter-core=5.5.0=pypi_0
jupyter-events=0.8.0=pypi_0
jupyter-lsp=2.2.0=pypi_0
jupyter-server=2.10.0=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyter_client=8.6.0=py311h06a4308_0
jupyter_core=5.5.0=py311h06a4308_0
jupyter_events=0.8.0=py311h06a4308_0
jupyter_server=2.10.0=py311h06a4308_0
jupyter_server_terminals=0.4.4=py311h06a4308_1
jupyterlab=4.0.8=pypi_0
jupyterlab-server=2.25.1=pypi_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.25.1=py311h06a4308_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
liac-arff=2.5.0=pyhd3eb1b0_1
libboost=1.82.0=h109eef0_2
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.8.1.2=0
libcurand=10.3.4.107=0
libcurl=8.5.0=h251f7ec_0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=hdbd6064_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=13.2.0=h807b86a_4
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnghttp2=1.57.0=h2d74bed_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libsodium=1.0.18=h7b6447c_0
libssh2=1.10.0=hdbd6064_2
libstdcxx-ng=13.2.0=h7e041cc_4
libtasn1=4.19.0=h5eee18b_0
libthrift=0.15.0=h1795dd8_2
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
minio=7.1.0=pyhd3eb1b0_0
mistune=2.0.4=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-fft=1.3.8=pypi_0
mkl-random=1.2.4=pypi_0
mkl-service=2.4.0=pypi_0
mkl_fft=1.3.8=py311h5eee18b_0
mkl_random=1.2.4=py311hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=pypi_0
nbclient=0.8.0=pypi_0
nbconvert=7.10.0=pypi_0
nbformat=5.9.2=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=pypi_0
notebook=7.0.7=pyhd8ed1ab_0
notebook-shim=0.2.3=pypi_0
numpy=1.26.3=pypi_0
numpy-base=1.26.3=py311hf175353_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openml=0.12.2=pyhd8ed1ab_0
openssl=3.0.12=h7f8727e_0
orc=1.7.4=hb3bc3d3_1
overrides=7.4.0=pypi_0
packaging=23.1=pypi_0
pandas=2.2.0=pypi_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pillow=10.0.1=pypi_0
pip=23.3.1=pypi_0
platformdirs=3.10.0=pypi_0
prometheus-client=0.14.1=pypi_0
prometheus_client=0.14.1=py311h06a4308_0
prompt-toolkit=3.0.43=pypi_0
prompt_toolkit=3.0.43=hd3eb1b0_0
psutil=5.9.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=11.0.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=pypi_0
pyopenssl=23.2.0=pypi_0
pysocks=1.7.1=pypi_0
python=3.11.7=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py311h06a4308_0
python-json-logger=2.0.7=pypi_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.11=2_cp311
pytorch=2.2.0=py3.11_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=pypi_0
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
referencing=0.30.2=pypi_0
requests=2.31.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rpds-py=0.10.6=pypi_0
scikit-learn=1.4.0=pypi_0
scipy=1.11.4=pypi_0
send2trash=1.8.2=pypi_0
setuptools=68.2.2=pypi_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.10=h6a678d5_1
sniffio=1.3.0=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.12=pypi_0
tbb=2021.8.0=hdb19cb5_0
terminado=0.17.1=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
torch=2.2.0=pypi_0
torchaudio=2.2.0=pypi_0
torchtriton=2.2.0=py311
torchvision=0.17.0=pypi_0
tornado=6.3.3=pypi_0
traitlets=5.7.1=pypi_0
triton=2.2.0=pypi_0
typing-extensions=4.9.0=pypi_0
typing_extensions=4.9.0=py311h06a4308_1
tzdata=2023d=h04d1e81_0
urllib3=1.26.18=pypi_0
utf8proc=2.6.1=h27cfd23_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=pypi_0
websocket-client=0.58.0=pypi_0
wheel=0.41.2=pypi_0
xmltodict=0.13.0=pypi_0
xz=5.4.5=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.5=h6a678d5_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
